frontend/requirements.txt
flet==0.23.1
httpx==0.25.1
python-dotenv==1.0.0
